Cleanup router and make line items controller
Prior to this change, order controller and view got really complex.

This change simplifies the routes as well as the controllers and views.
